+++ src/locale/useLocale.tsx
@@ -1,8 +1,46 @@
-import { useContext, useMemo } from 'react';
+import { createContext, useContext, useMemo } from 'react';
 
-import type { Locale, LocaleContextProps } from '../configProvider';
-import { LocaleContext } from '../configProvider';
 import defaultLocaleData from './zh-CN';
+
+export interface Locale {
+    locale: string;
+    BlockHeader: { expand: string; collapse: string };
+    Copy: { copied: string; copy: string };
+    Dropdown: { selectAll: string; resetText: string; okText: string };
+    Fullscreen: { exitFull: string; full: string };
+    GlobalLoading: { loading: string };
+    Input: {
+        case: string;
+        precise: string;
+        front: string;
+        tail: string;
+    };
+    LoadError: {
+        please: string;
+        get: string;
+        refresh: string;
+        title: string;
+    };
+    Modal: {
+        okText: string;
+        cancelText: string;
+    };
+    MxGraph: { newNode: string };
+    NotFound: {
+        description: string;
+    };
+    SpreadSheet: {
+        description: string;
+        copy: string;
+        copyAll: string;
+    };
+}
+
+export interface LocaleContextProps {
+    locale: Locale;
+}
+
+export const LocaleContext = createContext<LocaleContextProps>({ locale: {} as Locale });
 
 export type LocaleComponentName = keyof Locale;
